clean-up

+++ stm32/stm32/Src/lora.c
@@ -18,7 +18,7 @@ void setLR1110_LoRa_Packet_Type( const void* context, const uint8_t packet_type 
 
     cbuffer[0] = ( uint8_t )( LR1110_LORA_CMD_SET_PACKET_TYPE >> 8 );
     cbuffer[1] = ( uint8_t )( LR1110_LORA_CMD_SET_PACKET_TYPE >> 0 );
-    cbuffer[2] = packet_type;
+    cbuffer[2] = ( uint8_t ) packet_type;
 
     if ( lr1110_spi_write( context, cbuffer, LR1110_LORA_CMD_LENGTH_SET_PACKET_TYPE ) == LR1110_SPI_STATUS_OK ) {
         HAL_DBG_TRACE_MSG_COLOR("DONE\r\n", HAL_DBG_TRACE_COLOR_GREEN);
